Use pyunifi as it's maintained.
The old unifi.controller isn't maintained, and didn't work for me after I
upgraded my Unifi controller (to 5.5.20)

pyunifi is a maintained, more-or-less drop-in compatible replacement:
https://github.com/finish06/pyunifi

I added a --no_ssl_verify option which is passed to pyunifi, as the default is
to verify SSL certs, so if you use a self-signed cert internally the connection
would be rejected.

This fixes #6.